+++ setup.py
@@ -6,7 +6,7 @@ setup(
     description='SuiteSparse:GraphBLAS Python bindings.',
     packages=find_packages(),
     author='Michel Pelletier, James Kitchen, Erik Welch',
-    cffi_modules=["suitesparse/graphblas/build.py:ffibuilder"],
+    cffi_modules=["suitesparse_graphblas/build.py:ffibuilder"],
     install_requires=["cffi>=1.0.0"],
     setup_requires=["cffi>=1.0.0", "pytest-runner"],
     tests_require=["pytest"],
