+++ suitesparse_graphblas/clean_header.py
