+++ tests/test_package.py
@@ -1,4 +1,4 @@
-from suitesparse.graphblas import lib, ffi
+from suitesparse_graphblas import lib, ffi
 
 def test_matrix_existence():
     assert hasattr(lib, "GrB_Matrix_new")
